+++ cmd/utils/history.go
@@ -41,6 +41,19 @@ import "github.com/monax/relic"
 // release tagging script: ./scripts/tag_release.sh
 var History relic.ImmutableHistory = relic.NewHistory("fractal", "https://github.com/fractalplatform/fractal").
 	MustDeclareReleases(
+		"0.0.20 - 2019-06-12",
+		`### Fixed
+- [DOWNLOADER] fixed bug of find ancestor and use random station 
+- [BLOCKCHAIN] fixed blockchain irreversible number
+### Add
+- [DPOS] add thread test for rand vote candidate
+- [BLOCKCHAIN] add refuse bad block hashes
+- [BLOCKCHAIN] sync block with a specified block number
+`,
+		"0.0.19 - 2019-06-11",
+		`### Fixed
+- [ASSET] modify subasset decimals
+`,
 		"0.0.18 - 2019-06-06",
 		`### Fixed
 - [ACCOUNT] modify children check function